Moved ServiceInfo from spi to spi.impl.servicemanager
Although this class was part of a public API; there is no public API where this is being used.
So it isn't exposed in any way apart from being in the spi module. So I find it safe to
move inside the implementation.
