+++ hazelcast/src/main/java/com/hazelcast/spi/impl/servicemanager/ServiceInfo.java
@@ -14,7 +14,11 @@
  * limitations under the License.
  */
 
-package com.hazelcast.spi;
+package com.hazelcast.spi.impl.servicemanager;
+
+import com.hazelcast.spi.ConfigurableService;
+import com.hazelcast.spi.CoreService;
+import com.hazelcast.spi.ManagedService;
 
 /**
  * Contains the name of the service and the actual service.
